requirements.txt
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==2.1.5
PyMuPDF==1.24.11
python-docx==1.1.2
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
gunicorn
